pyproject.toml
[project]
name = "layers"
version = "3.0.19"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.8.5",
    "alembic==1.7.4",
    "cachetools==5.3.1",
    "fastapi==0.95.0",
    "minio==7.1.15",
    "psycopg2-binary==2.9.8",
    "pyjwt[crypto]==2.8.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "sqlalchemy[asyncio]==1.4.54",
    "sqlmodel==0.0.11",
    "uvicorn==0.15.0",
]

[dependency-groups]
dev = [
]
lint = [
    "ruff==0.11.4",
]
tests = [
    "httpx==0.23.3",
    "pytest-mock==3.10.0",
    "pytest==7.3.2",
    "testcontainers==4.10.0",
]
security = [
    "pip-audit==2.7.3",
]
migrations = [
    "alembic==1.7.4",
    "psycopg2-binary==2.9.8",
    "sqlmodel==0.0.11",
]
